+++ AdventOfCode.Test/Year2019/Day2Puzzle1Tests.cs
@@ -14,7 +14,7 @@ namespace Pontemonti.AdventOfCode.Test.Year2019
         {
             // Correct answer is 6327510
             int expectedResult = 6327510;
-            int actualResult = Day2Puzzle1.CalculateResult();
+            long actualResult = Day2Puzzle1.CalculateResult();
             Assert.AreEqual(expectedResult, actualResult);
         }
     }
