+++ AdventOfCode.Test/Year2019/Day7Puzzle1Tests.cs
@@ -16,7 +16,7 @@ namespace Pontemonti.AdventOfCode.Test.Year2019
         {
             // Correct answer is 75228
             int expectedResult = 75228;
-            (int actualResult, _) = Day7Puzzle1.GetResult(InputHelper.ReadIntegerCommaList(Day7Puzzle1.input).ToArray());
+            (long actualResult, _) = Day7Puzzle1.GetResult(InputHelper.ReadIntegerCommaList(Day7Puzzle1.input).ToArray());
             Assert.AreEqual(expectedResult, actualResult);
         }
 
@@ -55,7 +55,7 @@ namespace Pontemonti.AdventOfCode.Test.Year2019
         private void TestGetResult(string programString, int expectedOutputSignal)
         {
             int[] program = InputHelper.ReadIntegerCommaList(programString).ToArray();
-            (int actualOutputSignal, int[] actualPhaseSettings) = Day7Puzzle1.GetResult(program);
+            (long actualOutputSignal, int[] actualPhaseSettings) = Day7Puzzle1.GetResult(program);
             Assert.AreEqual(expectedOutputSignal, actualOutputSignal);
         }
     }
